# ⚡ Optimized KeyConverters