|      Method |     Mean |   Error |   StdDev |   Median |       Gen0 |      Gen1 |     Gen2 | Allocated |
|------------ |---------:|--------:|---------:|---------:|-----------:|----------:|---------:|----------:|
|     WithNew | 261.6 ms | 5.19 ms | 11.81 ms | 265.0 ms | 20500.0000 | 5500.0000 | 500.0000 |  129.7 MB |
|    WithPool | 384.0 ms | 7.76 ms | 22.87 ms | 390.3 ms | 17000.0000 | 6000.0000 |        - | 114.44 MB |
| WithBuilder | 265.6 ms | 5.30 ms | 14.05 ms | 269.0 ms | 15500.0000 | 5500.0000 | 500.0000 |  99.18 MB |
|  WithUnsafe | 147.7 ms | 2.90 ms |  6.38 ms | 147.2 ms | 10600.0000 | 3800.0000 | 400.0000 |  68.66 MB |